requirements.txt
flask
python-dotenv
requests
PyPDF2
gunicorn
pandas